PRO-2654 Updated to remove comments and additional information exposed because security is disabled for health and info.
