+++ src/main/java/uk/gov/hmcts/probate/services/persistence/ApplicationConfig.java
@@ -1,5 +1,12 @@
 package uk.gov.hmcts.probate.services.persistence;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import javax.sql.DataSource;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -11,13 +18,8 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
 import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurerAdapter;
 import org.springframework.jdbc.support.incrementer.PostgreSQLSequenceMaxValueIncrementer;
-import uk.gov.hmcts.probate.services.persistence.model.Submission;
 
-import javax.sql.DataSource;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
+import uk.gov.hmcts.probate.services.persistence.model.Submission;
 
 @EntityScan(basePackages = {"uk.gov.hmcts.probate.services.persistence.model"})
 @EnableJpaRepositories(basePackages = {"uk.gov.hmcts.probate.services.persistence.repository"})
