+++ src/main/resources/application.properties
@@ -5,6 +5,7 @@ probate.postgresql.user=probate
 probate.postgresql.password=password123
 probate.persistence.port=8282
 probate.persistence.show_sql=true
-
-# Allows gitCommitId to be publically exposed in the health end-point.
 management.security.enabled=false
+management.health.diskspace.enabled=false
+management.health.db.enabled=false
+management.info.git.enabled=false
